Bug #16072004 : GRANT IDENITFIED BY PASSWORD SUCCESSFUL FOR PAM/WINDOWS AUTHENTICATED USERS
When somebody tries to set a password for a user that's authenticated
via an external plugin the server will just silently ignore the change and 
won't do anything with the password. 
This is expected behavior : the external authentication is designed like 
this. 
But silently ignoring commands is unfriendly and counter-intuitive.

Fixed by adding a warning when this happens.
No test cases needed, since existing tests are already covering this
scenario.