Bug#16241773: RANGE OPTIMIZER CRASHES IN SEL_ARG::RB_INSERT()
In the range optimizer, an OR operation was performed between
a SEL_TREE with merges (two alternative UNION of indexes) and 
a simple SEL_TREE with only one range:

SEL_TREE1: idx1_fld>10 OR idx2_fld=10
           AND
           idx1_fld>20 OR idx2_fld=11
  OR
SEL_TREE2: idx1_fld=5

This OR operation was performed by imerge_list_or_tree(), which
intended to do this: 
  
1: tree_or((idx1_fld>10 OR idx2_fld=10), (idx1_fld=5))
2: tree_or((idx1_fld>20 OR idx2_fld=11), (idx1_fld=5))

But tree_or() may modify it's arguments, which is what happened
in this bug. Instead of doing the above, imerge_list_or_tree()
did this:

1: <same as above>
2: tree_or(<same_as_above>, (idx1_fld=5 OR idx1_fld>10))
                                        ^^^^^^^^^^^^^^

The fix is in imerge_list_or_tree(): Instead of calling 
tree_or() once and then starting to clone the second 
argument, we now make clones in all but the last time 
tree_or() is called. Thus, the SEL_TREE is now cloned before 
it may have been changed.