Bug#16310273 BUG WITH ALTER TABLE AUTO_INCREMENT
To be compatible with ALGORITHM=COPY, InnoDB has to mimic the
behaviour in ALGORITHM=INPLACE.  Instead of setting the AUTO_INCREMENT
value to the user-specified one, it has to set it to the maximum of
the user-specified value and MAX(auto_increment_column)+1.

commit_get_autoinc(): Get the maximum value.

rb#2005 approved by Kevin Lewis