Bug#16342793 MYSQL 5.6 FAILS TO BUILD WITH LIBWRAP SUPPORT ENABLED
Before this fix, building a server with HAVE_LIBWRAP failed.

The problem is that a variable renaming was missed, causing code to not
compile in sql/mysqld.cc

However, such failure should have been detected, and in fact was hidden by
another problem.

In the cmake files, the end user can specify WITH_LIBWRAP.

The Makefile then attempts to build a small test program to verify libwrap
is available, but silently fail and still produce code that does build,
with HAVE_LIBWRAP undefined, producing a binary that does not use libwrap.

For the cmake test to pass, both:
- the tcpd.h header file
- the libwrap library
need to be installed.
The header and the library are deployed in different packages on some OS
distributions, so that having the libwrap runtime alone is not sufficient.

As a result, builds where:
- the end user ask for WITH_LIBWRAP
- the libwrap library *is* installed
- the tcpd.h header is missing
would build, but not with #define HAVE_LIBWRAP,
producing a binary without failures and without libwrap.

This fix resolves two issues:
- the build break in sql/mysqld.cc has been resolved
- the configure.cmake now fails with a hard failure,
  returning an error, if the end user asked WITH_LIBWRAP
  but the makefile can not find a working libwrap.

The most reliable way to verify how a binary is build is to check:
ldd mysqld | grep libwrap
to make sure the library is called.

Verified manually that the build is now proper.