Patch for Bug#16076180 - ERROR 2027: MALFORMED PACKET, ERROR 2014: COMMANDS OUT OF SYNC.
The problem was that a failure in processing of an administrative statement
(CACHE INDEX) was not properly handled. In particular, the pending resultset
was not cancelled, which led to protocol error.

That could occur if the CACHE INDEX statement was incorrectly used inside
a stored program, where the thrown error was handled by an SP error handler.

The fix is to cancel active result set in case of error.