Bug#16284051 INNOCHECKSUM LINK FAILS WITH SYSTEM ZLIB Bug#68277 innochecksum link fails with system zlib
If cmake is invoked with -DWITH_SSL=</path/to/custom/openssl>
we try to link with the .a files rather than the .so files.
This functionality inadvertently changed behaviour for
 -DWITH_SSL=system
as well.
If the user has both /usr/lib/libcrypto.a and /usr/lib/libcrypto.so
we would pick the static library, and the link would fail with something like:
/usr/lib/gcc/i686-linux-gnu/4.7/../../../i386-linux-gnu/libcrypto.a(c_zlib.o): In function `bio_zlib_write':