Fix for Bug#16327534 ASSERT 'OTHER_TBLS_OK' FAILED IN PUSH_INDEX_COND                      ON SELECT + IN + JOIN WITH ICP
The failing assert was added to push_index_cond() as part of the fix
for Bug 15896009. The purpose of the assert was to ensure that when an
index condition is not pushed due to not including any fields from the
index, this only happens when we are allowed to include fields from
other tables of the join, ie. when other_tbls_ok is true.

The reason for the assert to fail is that there exist situations where
the query condition does not include any fields from any of the
tables. This can happen when we have a constant condition. One example
is that the query condition is just Item_int(1) (ie. always
true). This causes the assert to fail due to not referencing any
fields neither from the index nor from any other tables.

The fix for this problem is to extend the assert to also include the
case where the index condition is a constant condition. 
Note: in the version pushed to mysql-5.6 the assert is removed.