Bug#13944392 PERFORMANCE OVERHEAD IN MY_HASH_SORT_BIN
This fix is a performance improvement.

To sort strings according to character set and collation rules,
the character set code implements dedicated hash functions.

The implementation of these hash functions was too complex,
causing un necessary memory access with pointer indirections,
causing CPU overhead when computing a string hash.

This cause hash functions in general to appear in various
performance benchmarks, for example with oprofile,
or when benchmarking the performance schema.

The performance schema is affected in particular for:
- the table instrumentation (a LF_HASH by table schema + table name is
  used),
- the file io instrumentation (a LF_HASH by file name is used).

Any other code using a hash on a string is affected also.

The root cause of the problem is the code itself.

Typically, code looks like:

hash_function( ulong *nr /* INPUT and OUTPUT parameter */)
{
  /* loop in the string */
  for (...)
  {
    *nr = complex expression ( *nr );
  }
}

The problem is that there is no need to constantly read from and write to
the location of the nr pointer inside the loop, which causes a lot of memory
indirections.

The code has been changed to look like:

hash_function( ulong *nr /* INPUT and OUTPUT parameter */)
{
  ulong tmp;

  tmp = *nr; /* read INPUT parameter */

  /* loop in the string */
  for (...)
  {
    /* perform computation locally */
    tmp = complex expression ( tmp );
  }

  *nr = tmp; /* write OUTPUT parameter */
}

This change improves both clarity, and efficiency.

Every character set hash function is affected,
as all functions used the same coding pattern.