BUG#16035659 - CRASH WHEN SQL_MODE=NO_ENGINE_SUBSTITUTION,                STRICT_TRANS_TABLES IS SET
BACKGROUND:
An assertion failure in Diagnostics_area::set_ok_status() was
observed in MySQL Server when an InnoDB table is created having
violated the maximum key length limit when running in STRICT 
SQL MODE. Instead of throwing an error, the server crashes in
Debug Build.
On release builds set_ok_status() will just refuse to overwrite
the error status with OK, and the client will be sent the error 
message. So we will not get a crash and in fact get correct 
behavior in this case. 

ANALYSIS:
In THD::raise_condition() function, warning is converted to 
error when STRICT SQL MODE is enabled. The return value false
from this function indicates OK, but it has just set the error
status. There was no code to return true, if an error occured.

FIX:
As a fix for this bug, the code to return true has been added,
if an error is set when STRICT SQL_MODE is ON.