BUG#11766326 - LINUXTHREADS INFO IS OUTDATED
BACKGROUND:
In the Docs directory of the MySQL sourcecode there is a
linuxthreads.txt and a glibc-2.2.5.patch.
The first line of linuxthreads.txt already notes that it's
obsolete, but it's still in the sourcecode repository on
Launchpad as well as in internals Docs Directory.

FIX:
These two files have been removed from the internal Docs
Directory as the information in these files is obsolete 
and not required.