WL#6808 - InnoDB: Deprecate and remove innodb_mirrored_log_groups
Throw deprecation message in 5.6 and remove completely in 5.7

When innodb_mirrored_log_groups == 1, WARN and continue with the
following error message

"Innodb_mirrored_log_groups is an unimplemented feature and the variable
will be completely removed in a future version."

When innodb_mirrored_log_groups !=1, ABORT with the following error
message 

"Innodb_mirrored_log_groups is an unimplemented feature and the variable
will be completely removed in a future version. Using values other than 1
is not supported."

Approved by Marko. rb#1941