Bug#16066637-DROPPING A TABLE WITH NON-REGULAR CHARACTERS BREAKS REPLICATION
Problem:
========
1) Drop table queries are re-generated by server
before writing the events(queries) into binlog
for various reasons. If table name/db name contains
a non regular characters (like latin characters),
the generated query is wrong. Hence it breaks the
replication. 
2) In the edge case, when table name/db name contains
64 characters, server is throwing an assert
assert(M_TBLLEN < 128)

Analysis & Fix :
================
1) Parser reads the table name from the query and converts
it to standard charset and stores it in table_name variable.
When drop table query is regenerated with the same table_name
variable, it should be converted back to the original charset.
2) Latin character takes two bytes for each character. Limit
of the identifier is 64. Hence the assert should be less than 
or equal to 128 instead of less than 128.