Bug#16347450 GROUP_CONCAT UNSTABLE, CRASH IN ITEM_SUM::CLEAN_UP_AFTER_REMOVAL
Problem: Queries with ALL/ANY/SOME/IN subqueries with aggregation
functions in unnecessary clauses may cause the server to crash.

ALL/ANY/SOME/IN subqueries are simplified by removing clauses that
will not affect the result. This is done by calling
remove_redundant_subquery_clauses(). When such a clause contains an
aggregation function, Item_sum::clean_up_after_removal() is called to
unlink the item from the aggr_sel->inner_sum_func_list if such a link
exists.

Item_sum::clean_up_after_removal() assumes that the item is part of
the list if aggr_sel and aggr_sel->inner_sum_func_list are both
non-null. However, if the subquery has both an inner subquery with an
aggregation function and an aggregation function (not in an inner
subquery) that is to be removed, Item_sum::clean_up_after_removal()
will fail to identify the situation and will try to remove the item
from a list where it is not present. This causes it to dereference an
uninitialized next pointer.

Fix: Initialize Item_sum::next to NULL and use this to test for
membership in aggr_sel->inner_sum_func_list. The list is circularly
linked, so any element with next==NULL is not a member of the list.