Bug #16248315: mysql_secure_installation cannot verify newly set password with expired flag
the mysql command line binary would only turn on the 
MYSQL_OPT_CAN_HANDLE_EXPIRED_PASSWORD option when the client is 
interactive. This makes any scripting that tries to deal with password 
expiration impossible.
Added a new mysql command line/config file option to override setting 
this:
--connect-expired-password 
When this option is supplied it will cause mysql to always connect to the 
server, but be in "sandbox" mode if connecting with an expired password.
Updated the mysql_secure_installation script to use the new option.
test cases added.