Bug#16300620: VALGRIND FAILURE MYSQL-5.6/TRUNK IN               AUTH_SEC.KEY_VALUE_AUTH
This valgrind failure could occur if CREATE USER was logged in
the general query log and the 'old_passwords' system variable
was set to 2.

During query rewrite to prevent the password from being logged
in plain text, my_make_scrambled_password() was called. The
output of this function was stored in a char buffer. Afterwards
the contents of this char buffer was appended to the query string.

The problem was that we appended the number of characters
equal to the length of the plaintext password. So if the
scrabled password was shorter, we ended up appending uninitialized
bytes. This later caused a valgrind warning during logging.

This patch fixes the problem by writing "<secret>" to the query
log instead of the scrambled password for old_passwords = 2.

No test case added since this is covered by existing tests.