BUG#14781478 - CREATE SERVER: MEMORY LEAK WHEN STATEMENT FAILS!!
BACKGROUND:
If the `mysql`.`servers` table is readonly or missing, CREATE 
SERVER statements fail, and leak pool memory. This is because
we allocate memory for Foreign Server but we are not freeing it
when we are not able to insert it into mysql.servers table.
If CREATE SERVER statements are run over and over, memory will 
eventually be all consumed.

ANALYSIS:
The Problem is that we allocate memory in 
prepare_server_struct_for_insert() for a FOREIGN_SERVER 
object and then we insert it into mysql.servers table and 
servers cache.
If mysql.servers table does not exists or it cannot be opened,
server object is not stored in mysql.servers table and
the memory allocated for it is not freed which leads to the 
problem mentioned in the bug.
On further analysis, this problem is also present in update code
as well when we are going to update a record and mysql.servers
table cannot be opened. 

FIX:
As a fix for this bug, the code has been reorganized in such
a way that We do not allocate memory for server object before
we know that mysql.servers table does not exists or it cannot
be opened.
This reorganization is done in 4 steps:
1) Open mysql.servers table in (insert_server()/update_server()).
2) Insert/Update server record into mysql.servers in 
   (insert_server_record()/update_server_record()) functions using 
   the information provided from the parser in LEX_SERVER_OPTIONS struct.
3) Prepare a FOREIGN_SERVER object using 
   (prepare_server_struct_for_insert()/ prepare_server_struct_for_update())
   in (insert_server_record()/update_server_record()) functions respectively.
4) Insert/Update the record in servers cache using 
   (insert_server_record_into_cache()/update_server_record_in_cache())
   functions respectively.
This way we make sure that on failure, no memory leak occurs.
Also prepare_server_struct_for_update() is made similar to the function
prepare_server_struct_for_insert().