Fix for Bug#16328373 ASSERTION FAILED: ROWIDS_BUF < ROWIDS_BUF_END
The assert in DsMrr_impl::dsmrr_fill_buffer() fires due to the buffer
space to use for sort buffer is too small to store a single primary
key. The primary key is in this case large since it contains the
prefix of a BLOB. The buffer space to use for sort buffer is provided
by the BKA implementation. The query is run with a very small join
buffer of only 128 bytes. When BKA is used, the first part of the join
buffer is used for storing the fields from the first table while the
last part is used as sort buffer for the DS-MRR implementation.

In JOIN_CACHE::set_constants() there is already code for handling the
situation when the specified join buffer is too small. The code
adjusts the size of the join buffer if the size specified by
join_buffer_size is so small that it can not store a single
record. This code does not take into account that there needs to be
room in the join buffer for the DS-MRR sort buffer when BKA or
BKA-unique is used.

The fix for this problem is to extend JOIN_CACHE::set_constants() so
that when it computes the minimum size the join buffer can have, it
will include space for a minimal DS-MRR sort buffer.