Bug#16341673: ASSERTION FAILED: THD->TRANSACTION.XID_STATE.XID.IS_NULL(),               FILE TRANSACTION.CC,
The triggered assert checks that the XA transaction state is
properly reset before trying to start an XA transaction.
The problem was that it could be triggered if a statement had
failed in autocommit mode right before XA START was issued.

The bug was that rollback of statement transactions in auto
commit mode did not call transaction.cleanup(). This function,
among other things, resets XID state. And reset XID state is
checked for during XA START.

This patch fixes the problem by reverting to 5.5 behavior and
calling transaction.cleanup() for rollback in autocommit mode.

This was a regression introduced in 5.6 by WL#5223.