Fix for bug16402744: CRASH DURING TC EXECUTING CONTINUEB(ZSTART_FRAG_SCAN)
A crash was caused by startFragScanLab() when it incorectly assumed it
was about to start a fragment scan for the last fragment and thus
released the memory sections holding the Attr- and KeyInfo for the 
entire scan.

Any outstanding CONTINUEB(ZSTART_FRAG_SCAN) would then later crash due
to these released sections.

This fix ensures that no more sendDihGetNodesReq() are sent when there
can be outstanding CONTINUEBs. When CONTINUEB completes, sendDihGetNodesReq()
will be resumed. This change ensures that all fragments will be processed
in sequence.