BUG#16062608: REPLICATION COMMANDS IMPLICIT COMMIT AND SET GTID_NEXT="UUID:X"
When @@SESSION.GTID_NEXT is set to "UUID:X" and a statement that
implicit commits but does not log changes is executed, like START SLAVE,
that GTID is logged as a empty transaction, when on the same scenario
with gtid-mode=OFF nothing is logged.

This situation is not desirable since it may cause inconsistency
problems:
  1) DBA is performing a restore or recovery;
  2) She/he is applying some empty transactions on slave to avoid
     problematic transactions from master;
  3) Made a mistake and set a not need to skip GTID and didn't revert
     it.
  4) Start slave.
  5) That logged GTID, which will cause that the transaction from
     master will be skipped, will eventually cause inconsistency
     problems.

To avoid that kind of mistakes, when a command that causes a implicit
commit but does not change data is executed when @@SESSION.GTID_NEXT is
different from AUTOMATIC, a error is thrown saying that this command
is not allowed.