Fix for bug#16199028 INCORRECTLY REJECT TO INCLUDE TABLE AS PART OF 'PUSHED JOIN'
Updated fix according to comments from reviewer.
Also slightly changed how the 'common_ancestor' is calculated as
Jan Wedvik found another testcase during review which should have
been pushed. Added this testcase as well.
      
..................

Including a table a part of a pushed join should be rejected if there
are outer joined tables inbetween the table to be included, and the tables
it is joined with. Including such a table in the pushed join would have
introduced implicit join dependencies with the outer joined table.
This effectively would have changed the semantics of the query, and 
as such would have caused incorrect results to be returned.
      
However, the implementation of the above restriction incorrectly
checked for any such outer joined tables by checking the join type 
against the root of the pushed query. The correct check would have 
been to use the common ancestor of the tables we joins with.
      
This fix calculates this common ancestor and use that as described above.
Furthermore it enhances the 'explain extended' output which explain why
a table potentially rejected by this test cant be included in the pushed join.
      
Also add some more comments to this region of the code and fixed a few typos.