Prepare MTR test for features deprecated in mysql-5.6-cluster-7.3
This is a collection of patches pushed to mysql-trunk-cluster 
by Jan Wedvik early september 2012 in order to get the daily 'big'
tests running. We assumed that this 'trunk' work would eventually 
be merged into 5.6-cluster-7.3, but they apperently didn't reach
a cutoff date for changed merge policy.

Thus this fix to rectify that.