Bug #16043216 DROP DATABASE: ASSERTION: !TABLE->CAN_BE_EVICTED || TABLE->IBD_FILE_MISSING
Problem 1:

When a drop dabase is done the SQL layer do the following steps:

1.  Drop each table individually from InnoDB.
2.  Drop database from InnoDB

But, if the frm file of a particular table is missing, step (1)
will not happen.  Because of this, in step (2), the can_be_evicted
flag of the table will be TRUE.   In step (2), we assert that
can_be_evicted flag of the table will be FALSE.

Solution:

Under certain situations, the frm file of a table can be missing.
During drop database, if the can_be_evicted flag is TRUE, then
issue a warning, but proceed with the operation.  There is no need
to assert.

Problem 2:

When an ALTER TABLE command is issued against a table that contains
an FTS index to add a foreign key constraint, and if this alter fails,
then proper cleanup was not happening.  The FTS index tables that was
created was not cleaned up, but were left as orphans.  Because of this
a subsequent drop database statement was failing.  

Proper cleanup of the FTS index tables was not happening because the
necessary FTS index information was not loaded when the failure happened.
The error handling code didn't have access to the FTS index information.

Solution:

The FTS index information is loaded before attempting to add the foreign
key constraint.  This ensures that proper cleanup is done when alter
table fails to add a foreign key constraint.

rb#2119 approved by Marko and Jimmy.