Bug#16429216: SP_NOTEMBEDDED IS UNSTABLE ON PB2
sp_notembedded.test was unstable because of the regression test for Bug#47736.
This regression test is supposed to kill a select from a view while the view
was processing a function. This was implemented by having a function that executed
sleep(60). The problem was that there was no synchronization ensuring that sleep(60)
was actually executing when KILL was issued.

This patch fixes the problem by reimplementing the stored function to use
user-level locks for synchronization. The patch also removes an extra
connection that was not needed.