Bug#16296837-INNODB: DELETE DEAD CODE PROTECTED BY MACRO UNIV_LOG_ARCHIVE
The code protected by macro UNIV_LOG_ARCHIVE has not been compiled for years
and probably would not work anymore.  It was never completed, with asserts
that would be hit regularly.  The plan for the future is to rewrite the REDO
log to combine it with the binlog, which is already archiveable.