Bug #11763660 - PLEASE ASSERT INSTEAD OF CALLING EXIT(1)
This patch replaces exit(1); with ut_error; in most places.
There are a few calls to exit(3) that are left as is because the code
seems to require it.

In addition, whenever ut_error; is preceeded by a log message, use
ib_logf(IB_LOG_LEVEL_FATAL,...) instead of fprintf(stderr,...)

Code approved by Marko in http://rb.no.oracle.com/rb/r/1937/