BUG#16367039 - LOCK TIMEOUT IN UPDATE SUBQUERY CAUSES ASSERT
ANALYSIS :
When an UPDATE statement containing a nested SELECT statement is being executed
and another connection happens to hold locks to the rows being accessed by this thread 
then optimize_cond() in mysql_update() fails due to a lock wait timeout and the
Diagnostic Area status is set to DA_ERROR.

However this DA status is not checked and the execution continues, subsequently call to
info.read_record() returns -1 which is a Success condition stating that all records have been
processed. This results in a call to my_ok() which causes an ASSERT because the DA has already
been set to DA_ERROR. 

In release build however Diagnostics_area::set_ok_status() has explicit code to which set DA_OK 
only if the DA state has not already been set, therefore the behavior remains consistent with debug 
builds.

FIX :
The DA status is checked for an error status and if it has been set, the execution of the statement
aborted.