Bug#14749800 - ERROR 1571 (HY000): ERROR DURING STARTING/                STOPPING OF THE SCHEDULER. ERROR CODE.
Bug#14781590 - SET GLOBAL EVENT_SCHEDULER = 1 SHOULD BE 
               SET TO 0 ON ERROR

Analysis:
-----------
In cases when we failed to create thread for event scheduler,
event execution or simply for handling new connection nothing
was logged into error log. 

Also system variable EVENT_SCHEDULER was not reset to OFF when
we failed to create event scheduler thread. This caused wrong
impression that event scheduler was running in cases when it is
not.

Fix:
-----------
Added writing message with errno to error log in the above cases.
Also adjusted other messages which were written to log when we
failed to create thread in other cases to contain an errno.
Also reset system variable EVENT_SCHEDULER to OFF on failure to
event scheduler create thread.

But logging error for thread creation failure for new connection
may spam error log.Used log throttling mechanism to avoid spamming
error log in such a situation. To do this kept generic log-throttling
functionality in Log_throttle class, moved functionality specific
for slow-log to Slow_log_throttle class. Added new 
Error_log_throttle class to be used for throttling of writing
to error log.