Fix for Bug#14554519 SHOULD NOT SET SOCKET OPTIONS 'SO_RCVBUF' AND 'SO_SNDBUF' WHEN NOT CONFIGURED
This fix will only conditionally set the socket options SO_RCVBUF, and SO_SNDBUF if they
was specified in config.ini. If these config parameters was not specified, the OS is free
to choose the size on its own.
      
The default value of resp. 70080/71540 for the Rcv/snd buffer is also removed, and
this change has to be reflected in the MySQL Cluster documentation.