BUG#16366881 - LOCK TIMEOUT DURING FILESORT CAUSES ASSERT
ANALYSIS :
When a query is being executed and another connection happens to hold
locks being accessed by this thread, then multi_range_read_next()
called by the range optimizer could fail with HA_ERR_LOCK_DEADLOCK or
HA_ERR_LOCK_WAIT_TIMEOUT. However, this was not detected by
find_all_keys() and no error was reported to the Diagnostics Area.
Hence an assertion would be triggered in filesort(). On release
builds the consequence of this bug was that the ER_FILSORT_ABORT
error was sent to the client rather than the real error reported
by the storage engine. 
      
FIX :
If call to select->quick->reset() in find_all_keys() fails then an error is reported
using file->print_error() to set DA to a proper state.