BUG#16366994 - DBUG_ASSERT HIT WHILE EXECUTING UPDATE OR DELETE STATEMENT
ANALYSIS :
When an UPDATE or DELETE statement is being executed and another
connection happens to hold locks being accessed by this thread,
then multi_range_read_next() called by the range optimizer could
fail with HA_ERR_LOCK_DEADLOCK or HA_ERR_LOCK_WAIT_TIMEOUT. However,
this was not detected by mysql_update() and no error was reported
to the Diagnostics Area. Hence an assertion would be triggered
in Protocol::end_statement() when trying to send OK or ERROR to
the client. On release builds the bug caused OK to be sent to the
client for a statement that had in fact failed.

FIX :
If call to select->quick->reset() in mysql_update() or mysql_delete() fails then
an error is reported using file->print_error() to set DA to a proper state.