Bug#16439245 SERVER STUCK IN WAITING FOR SLAVE WORKERS TO FREE PENDING EVENTS
According to the bug description the Coordinator thread is hanging in
a while loop trying to catch decreased value of buffer size of 
queued for execution events.
Since Workers are idling per description that leaves the only possibility 
the current being queued event size exceeds the max allowed for the buffer.

Fixed with checking the being scheduled event size to error out by Coordinator
when it's bigger than the max.