Bug#16164885 - ASSERTION FAILED: TAB->TYPE == JT_REF,                FILE SQL_EXECUTOR.CC, LINE 3292
In 5.6 an assert was added in create_sort_index() under the assumption
that filesort was never used for tables with EQ_REF access.
However, the test case added here shows that this is currently not
true for subqueries.  Sorting is not really necessary in this case,
since EQ_REF means that there will be only one row to sort, and a
bug (16452392) has been filed for this issue.

For now, the assert is extended to also allow EQ_REF access.
Note that SQL_BIG_RESULT is required to hit this assert in the
supplied test case since this will force MySQL to create an sort
index.  (See comment on SQL_BIG_RESULT in JOIN::optimize)