Fix for Bug#16491986 AVOID EXTRA READTUPLE() WHEN UPDATE IS EXECUTED AS DELETE+REINSERT
When an update operation changes (part of) the primary key, the
operation is executed as a delete + reinsert.

If the initial read operation didn't retrieve all column values
required by the reinsert, another readTuple() was required
as part of ::ndb_pk_update_row().

This fix ensures that all required column values are included
in the first readTuple such that we saves a roundtrip.