Bug#15987941: REMOVE HAVE_NDB_BINLOG IN RPL CODE
Inserted in the current effort to clean the MySQL code of unused
ifdefs, there are several usages of the REMOVE HAVE_NDB_BINLOG flag in
the replication code that need to be removed.

Trough the removal/substitution of unused code, or the flag itself, the
several occurrences of this flag in non-NDB code are now gone.